Move test helper to a folder
Tests will be built using esbuild and this file will be replaced with a Deno compatible version when testing.

By compiling using esbuild migrating to Typescript should be easier.
